sound_generators.rs を大幅に書き直し GeneratorUnit毎に位相の管理を行うことで周波数変換処理を簡素化 周波数変換＋ミキシングを192kHzで行い最後に設定サンプリングレートにダウンコンバート 周波数変換とゲインコントロールの演算を整数化

+++ src/sound_manager/envelope_tbl.rs
@@ -1,4 +1,4 @@
-pub const ENVELOPE_TBL: &[&[u8]] = &[
+pub const ENVELOPE_TBL: &[&[i32]] = &[
     &[0x0f, 0x10],
     &[0x0c, 0x10],
     &[0x0a, 0x10],
